requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pyyaml>=6.0
lxml>=5.0.0
tzdata>=2024.1
